Delete category

+++ models/Category.php
@@ -69,7 +69,23 @@ class Category {
         return false;
     }
 
+    private function exists()
+    {
+        $query = "SELECT COUNT(id) as NUM FROM {$this->table} WHERE id = :id";
+        $statement = $this->conn->prepare($query);
+
+        $this->id = (int) htmlspecialchars(strip_tags($this->id));
+        $statement->bindParam(':id', $this->id);
+
+        $statement->execute();
+        return (bool) ((int) ($statement->fetch(PDO::FETCH_ASSOC))['NUM']);
+    }
+
     public function delete() {
+        if (!$this->exists()) {
+            return false;
+        }
+
         $query = "DELETE FROM {$this->table} WHERE id = :id";
         $statement = $this->conn->prepare($query);
